# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: API сервиса аутентификации с Redis и JWT. Поддерживает роли пользователей
    и администраторов.
  title: Device emission Service API
  version: "1.0"
paths:
...
  /api/users/:id:
    get:
      description: Возвращает информацию о пользователе по его ID (логин и роль)
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с данными пользователя
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID пользователя
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить информацию о пользователе
      tags:
      - Управление пользователями
    put:
      consumes:
      - application/json
      description: Обновляет логин и/или пароль пользователя по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно обновлен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный ID или тело запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при обновлении пользователя
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить данные пользователя
      tags:
      - Управление пользователями
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"